requirements.txt
ffmpeg==1.4
moviepy==1.0.3
imageio==2.24.0
google-auth==2.16.0
google-crc32c==1.5.0
imageio-ffmpeg==0.4.8
google-api-core==2.11.0
google-cloud-core==2.3.2
google-cloud-speech==2.17.0
google-cloud-storage==2.7.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0